setup-environment-internal: do not update the working tree during setup
it seems preferable to let the user sync her repo whenever it is the right time
to do, instead of doing it by default when doing the init.

Also, there is no specific reason to create a 'work' topic branch, on behalf of
the user, let them deal with that.

Change-Id: I524f7e93e6e8f2588fdd2a73742da1409840a999
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>