setup-environment-internal: add support for BSP's EULA
Some BSP might need EULA to be accepted, such as for example the BSP from
meta-qcom. This is a generic implementation of the EULA mechanism. If a BSP
needs a EULA, it needs to provide the EULA text file as conf/EULA/$MACHINE. In
which case the text will be presented to the user, who can decide to accept it
or not.

When accepted, the variable ACCEPT_EULA_$MACHINE is set to '1' in auto.conf, and
it can be used by the BSP layer.

For automation and CI, it is possible to preset the value of
ACCEPT_EULA_$MACHINE, by setting EULA_$MACHINE in the environment, before
calling setup-environment scripts.

Change-Id: I9fead8a8390bdee20d56cb3d22fa2ba3f42f2371
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>